+++ src/mus2mid.cpp
@@ -390,7 +390,7 @@ static int GetMIDIChannel(int mus_channel, MEMFILE *midioutput)
             // First time using the channel, send an "all notes off"
             // event. This fixes "The D_DDTBLU disease" described here:
             // https://www.doomworld.com/vb/source-ports/66802-the
-            WriteChangeController_Valueless(channel_map[mus_channel], 0x7b,
+            WriteChangeController_Valueless(static_cast<uint8_t>(channel_map[mus_channel]), 0x7b,
                 midioutput);
         }
 
@@ -520,7 +520,7 @@ bool mus2mid(MEMFILE *musinput, MEMFILE *midioutput)
                     return true;
                 }
 
-                if (WriteReleaseKey(channel, key, midioutput))
+                if (WriteReleaseKey(static_cast<uint8_t>(channel), key, midioutput))
                 {
                     return true;
                 }
@@ -543,7 +543,7 @@ bool mus2mid(MEMFILE *musinput, MEMFILE *midioutput)
                     channelvelocities[channel] &= 0x7F;
                 }
 
-                if (WritePressKey(channel, key,
+                if (WritePressKey(static_cast<uint8_t>(channel), key,
                         channelvelocities[channel], midioutput))
                 {
                     return true;
@@ -556,7 +556,7 @@ bool mus2mid(MEMFILE *musinput, MEMFILE *midioutput)
                 {
                     break;
                 }
-                if (WritePitchWheel(channel, static_cast<short>(key * 64), midioutput))
+                if (WritePitchWheel(static_cast<uint8_t>(channel), static_cast<short>(key * 64), midioutput))
                 {
                     return true;
                 }
@@ -573,7 +573,7 @@ bool mus2mid(MEMFILE *musinput, MEMFILE *midioutput)
                     return true;
                 }
 
-                if (WriteChangeController_Valueless(channel,
+                if (WriteChangeController_Valueless(static_cast<uint8_t>(channel),
                         controller_map[controllernumber],
                         midioutput))
                 {
@@ -608,7 +608,8 @@ bool mus2mid(MEMFILE *musinput, MEMFILE *midioutput)
                         return true;
                     }
 
-                    if (WriteChangeController_Valued(channel,
+                    if (WriteChangeController_Valued(
+                            static_cast<uint8_t>(channel),
                             controller_map[controllernumber],
                             controllervalue,
                             midioutput))
@@ -702,7 +703,7 @@ int main(int argc, char *argv[])
 
     infile_len = M_ReadFile(argv[1], &infile);
 
-    src = mem_fopen_read(infile, infile_len);
+    src = mem_fopen_read(infile, static_cast<size_t>(infile_len));
     dst = mem_fopen_write();
 
     if (mus2mid(src, dst))
